+++ keyboards/handwired/onekey/teensy_32/rules.mk
@@ -1,5 +1,8 @@
 # MCU name
 MCU = MK20DX256
 
+# Bootloader selection
+BOOTLOADER = halfkay
+
 # Enter lower-power sleep mode when on the ChibiOS idle thread
 OPT_DEFS += -DCORTEX_ENABLE_WFI_IDLE=TRUE
