+++ keyboards/handwired/onekey/teensy_lc/rules.mk
@@ -2,5 +2,8 @@
 MCU = MKL26Z64
 USE_CHIBIOS_CONTRIB = yes
 
+# Bootloader selection
+BOOTLOADER = halfkay
+
 # Enter lower-power sleep mode when on the ChibiOS idle thread
 OPT_DEFS += -DCORTEX_ENABLE_WFI_IDLE=TRUE
