+++ keyboards/mechlovin/olly/jf/rules.mk
@@ -5,7 +5,7 @@ MCU = atmega32a
 F_CPU = 16000000
 
 # Bootloader selection
-BOOTLOADER = USBasp
+BOOTLOADER = usbasploader
 
 # Build Options
 #   change yes to no to disable
