+++ keyboards/rocketboard_16/rules.mk
@@ -8,7 +8,7 @@ MCU = STM32F103
 MCU_LDSCRIPT = STM32F103xB_stm32duino_bootloader
 OPT_DEFS += -DBOOTLOADER_STM32DUINO
 BOARD = STM32_F103_STM32DUINO
-STM32_BOOTLOADER_ADDRESS = 0x80000000
+BOOTLOADER_TYPE = stm32duino
 DFU_ARGS = -d 1EAF:0003 -a 2 -R
 DFU_SUFFIX_ARGS = -v 1EAF -p 0003
