+++ tmk_core/protocol/arm_atsam/arm_atsam_protocol.h
@@ -19,7 +19,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define _ARM_ATSAM_PROTOCOL_H_
 
 #include "samd51j18a.h"
-#include "md_bootloader.h"
 
 #include "timer.h"
 #include "d51_util.h"
